start organizing for user supplied variable coders
